# backend/ms-orders/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentación Swagger para ms-orders",
        "title": "Orders API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/",
    "paths": {
        "/pedidos": {
            "get": {
                "description": "Devuelve todos los pedidos registrados con sus ítems",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Lista todos los pedidos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Pedido"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un pedido con los datos enviados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Crea un nuevo pedido",
                "parameters": [
                    {
                        "description": "Datos del pedido",
                        "name": "pedido",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Pedido"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Pedido"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pedidos/customers/{id}": {
            "get": {
                "description": "Devuelve los pedidos asociados a un cliente específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedidos"
                ],
                "summary": "Lista pedidos por cliente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Pedido"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.ItemPedido": {
            "type": "object",
            "properties": {
                "cantidad": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "pedido_id": {
                    "type": "integer"
                },
                "plato_id": {
                    "type": "string"
                },
                "precio": {
                    "type": "number"
                }
            }
        },
        "models.Pedido": {
            "type": "object",
            "properties": {
                "cliente_id": {
                    "type": "integer"
                },
                "estado": {
                    "type": "string"
                },
                "fecha": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ItemPedido"
                    }
                },
                "notas": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                }
            }
        }
    }
}